Azure's mqtt_client_connect() conflicts with a function from lwIP
Workaround this problem by prefixing the name of the function, which
results in azure_mqtt_client_connect(). This is redefined only
internally, this change is not visible outside of Azure, especially it
is NOT visible to the main application.

It is advisable to also redefine the name of the lwIP function, by
adding this in your main CMakeLists.txt:

target_compile_definitions(lwipallapps PRIVATE
		mqtt_client_connect=lwIP_mqtt_client_connect)

If you want to use one of these functions in your application directly,
also add this for the lwIP version:

target_compile_definitions(<application-target-name> PRIVATE
		mqtt_client_connect=lwIP_mqtt_client_connect)

or this for Azure version:

target_compile_definitions(<application-target-name> PRIVATE
		mqtt_client_connect=azure_mqtt_client_connect)